Modularize meeting capture context
Separate chat-scope validation and stable capture identity helpers behind the meeting accessibility facade without changing detection behavior.
